+++ testapp/manage.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python
+import logging
 import os
 import sys
 
@@ -11,6 +12,7 @@ if __name__ == "__main__":
     from djangae.core.management import execute_from_command_line, test_execute_from_command_line
 
     if "test" in sys.argv:
+        logging.disable(logging.CRITICAL)
         test_execute_from_command_line(sys.argv)
     else:
         execute_from_command_line(sys.argv)