+++ CHANGELOG.md
@@ -15,6 +15,8 @@
  - Fixed performance issues and bugs in the Djangae core paginator
  - Fix several issues with the test sandbox
  - Initialize the app_identity stub in the test sandbox
+ - Replace `print()` statements with `logging.debug()` in all unittests
+ - Silence stdout output during testing
  - Logging output silenced during `manage.py test` execution
 
 ## v0.9.10