+++ CHANGELOG.md
@@ -10,6 +10,7 @@
  in migrations.
 - Improve the approx SQL representation of Datastore commands (update, delete etc.)
 - Default value for failure_behaviour in `process_task_queues` is now `RAISE_ERROR`. Tasks will no longer fail silently when processed using this method in unit tests.
+- Add djangae.compat to handle SDK structural changes
 
 ### Bug fixes:
 
@@ -21,6 +22,7 @@
  - Logging output silenced during `manage.py test` execution
  - Fix management command `--help` output
  - Create .editorconfig to ensure basic editor settings are consistent between users
+ - Fix import error in SDK 1.9.60 
 
 ## v0.9.10